# apps/api/openapi.generated.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Soonr API",
    "description": "Fastify backend for Soonr application routes, interactive docs, and generated client contracts.",
    "version": "0.1.0"
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
...
  },
  "paths": {
...
    "/titles": {
      "get": {
        "summary": "Search titles",
        "tags": [
          "titles"
        ],
        "parameters": [
          {
            "schema": {
              "minLength": 2,
              "type": "string"
            },
            "in": "query",
            "name": "query",
            "required": true,
            "description": "Search query text."
          },
          {
            "schema": {
              "type": "string"
            },
            "in": "query",
            "name": "page",
            "required": false,
            "description": "Optional positive integer page number."
          },
          {
            "schema": {
              "type": "string"
            },
            "in": "query",
            "name": "limit",
            "required": false,
            "description": "Optional positive integer page size."
          },
          {
            "schema": {
              "type": "string"
            },
            "in": "query",
            "name": "forceRefresh",
            "required": false,
            "description": "Optional truthy flag: 1, true, or yes."
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "query",
                    "results",
                    "totalCount",
                    "page",
                    "limit",
                    "hasMore"
                  ],
                  "properties": {
                    "query": {
                      "type": "string"
                    },
                    "results": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "required": [
                          "id",
                          "kind",
                          "source",
                          "externalId",
                          "slug",
                          "name",
                          "coverImageUrl",
                          "earliestReleaseDate",
                          "platforms",
                          "rawgRating",
                          "rawgRatingsCount",
                          "rawgMetacritic",
                          "rawgAdded",
                          "rawgReviewsCount",
                          "rawgSuggestionsCount",
                          "rawgRatingTop"
                        ],
                        "properties": {
                          "id": {
                            "type": "string"
                          },
                          "kind": {
                            "type": "string",
                            "enum": [
                              "game"
                            ]
                          },
                          "source": {
                            "type": "string",
                            "enum": [
                              "rawg"
                            ]
                          },
                          "externalId": {
                            "type": "string"
                          },
                          "slug": {
                            "type": "string"
                          },
                          "name": {
                            "type": "string"
                          },
                          "coverImageUrl": {
                            "anyOf": [
                              {
                                "type": "string"
                              },
                              {
                                "type": "null"
                              }
                            ]
                          },
                          "earliestReleaseDate": {
                            "anyOf": [
                              {
                                "type": "string"
                              },
                              {
                                "type": "null"
                              }
                            ]
                          },
                          "platforms": {
                            "type": "array",
                            "items": {
                              "type": "object",
                              "required": [
                                "id",
                                "name"
                              ],
                              "properties": {
                                "id": {
                                  "type": "string"
                                },
                                "name": {
                                  "type": "string"
                                }
                              }
                            }
                          },
                          "rawgRating": {
                            "anyOf": [
                              {
                                "type": "number"
                              },
                              {
                                "type": "null"
                              }
                            ]
                          },
                          "rawgRatingsCount": {
                            "anyOf": [
                              {
                                "type": "number"
                              },
                              {
                                "type": "null"
                              }
                            ]
                          },
                          "rawgMetacritic": {
                            "anyOf": [
                              {
                                "type": "number"
                              },
                              {
                                "type": "null"
                              }
                            ]
                          },
                          "rawgAdded": {
                            "anyOf": [
                              {
                                "type": "number"
                              },
                              {
                                "type": "null"
                              }
                            ]
                          },
                          "rawgReviewsCount": {
                            "anyOf": [
                              {
                                "type": "number"
                              },
                              {
                                "type": "null"
                              }
                            ]
                          },
                          "rawgSuggestionsCount": {
                            "anyOf": [
                              {
                                "type": "number"
                              },
                              {
                                "type": "null"
                              }
                            ]
                          },
                          "rawgRatingTop": {
                            "anyOf": [
                              {
                                "type": "number"
                              },
                              {
                                "type": "null"
                              }
                            ]
                          }
                        }
                      }
                    },
                    "totalCount": {
                      "type": "number"
                    },
                    "page": {
                      "type": "number"
                    },
                    "limit": {
                      "type": "number"
                    },
                    "hasMore": {
                      "type": "boolean"
                    },
                    "servedBy": {
                      "anyOf": [
                        {
                          "type": "string",
                          "enum": [
                            "local-cache"
                          ]
                        },
                        {
                          "type": "string",
                          "enum": [
                            "rawg-refresh"
                          ]
                        }
                      ]
                    },
                    "decisionReason": {
                      "anyOf": [
                        {
                          "type": "string",
                          "enum": [
                            "local_sufficient"
                          ]
                        },
                        {
                          "type": "string",
                          "enum": [
                            "sparse_broad_local"
                          ]
                        },
                        {
                          "type": "string",
                          "enum": [
                            "forced_refresh"
                          ]
                        },
                        {
                          "type": "string",
                          "enum": [
                            "provider_missing_key"
                          ]
                        },
                        {
                          "type": "string",
                          "enum": [
                            "provider_fetch_failed"
                          ]
                        },
                        {
                          "type": "string",
                          "enum": [
                            "provider_used"
                          ]
                        }
                      ]
                    },
                    "providerUsedTrigger": {
                      "anyOf": [
                        {
                          "type": "string",
                          "enum": [
                            "coverage"
                          ]
                        },
                        {
                          "type": "string",
                          "enum": [
                            "freshness"
                          ]
                        },
                        {
                          "type": "string",
                          "enum": [
                            "coverage_and_freshness"
                          ]
                        }
                      ]
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "error"
                  ],
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "error"
                  ],
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/titles/{titleId}": {
      "get": {
        "summary": "Get title details",
        "tags": [
          "titles"
        ],
        "parameters": [
          {
            "schema": {
              "minLength": 1,
              "type": "string"
            },
            "in": "path",
            "name": "titleId",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "details",
                    "isInWatchlist"
                  ],
                  "properties": {
                    "details": {
                      "type": "object",
                      "required": [
                        "id",
                        "kind",
                        "source",
                        "externalId",
                        "slug",
                        "name",
                        "coverImageUrl",
                        "earliestReleaseDate",
                        "platforms",
                        "rawgRating",
                        "rawgRatingsCount",
                        "rawgMetacritic",
                        "rawgAdded",
                        "rawgReviewsCount",
                        "rawgSuggestionsCount",
                        "rawgRatingTop",
                        "description",
                        "genres",
                        "developers",
                        "publishers",
                        "websiteUrl",
                        "releases"
                      ],
                      "properties": {
                        "id": {
                          "type": "string"
                        },
                        "kind": {
                          "type": "string",
                          "enum": [
                            "game"
                          ]
                        },
                        "source": {
                          "type": "string",
                          "enum": [
                            "rawg"
                          ]
                        },
                        "externalId": {
                          "type": "string"
                        },
                        "slug": {
                          "type": "string"
                        },
                        "name": {
                          "type": "string"
                        },
                        "coverImageUrl": {
                          "anyOf": [
                            {
                              "type": "string"
                            },
                            {
                              "type": "null"
                            }
                          ]
                        },
                        "earliestReleaseDate": {
                          "anyOf": [
                            {
                              "type": "string"
                            },
                            {
                              "type": "null"
                            }
                          ]
                        },
                        "platforms": {
                          "type": "array",
                          "items": {
                            "type": "object",
                            "required": [
                              "id",
                              "name"
                            ],
                            "properties": {
                              "id": {
                                "type": "string"
                              },
                              "name": {
                                "type": "string"
                              }
                            }
                          }
                        },
                        "rawgRating": {
                          "anyOf": [
                            {
                              "type": "number"
                            },
                            {
                              "type": "null"
                            }
                          ]
                        },
                        "rawgRatingsCount": {
                          "anyOf": [
                            {
                              "type": "number"
                            },
                            {
                              "type": "null"
                            }
                          ]
                        },
                        "rawgMetacritic": {
                          "anyOf": [
                            {
                              "type": "number"
                            },
                            {
                              "type": "null"
                            }
                          ]
                        },
                        "rawgAdded": {
                          "anyOf": [
                            {
                              "type": "number"
                            },
                            {
                              "type": "null"
                            }
                          ]
                        },
                        "rawgReviewsCount": {
                          "anyOf": [
                            {
                              "type": "number"
                            },
                            {
                              "type": "null"
                            }
                          ]
                        },
                        "rawgSuggestionsCount": {
                          "anyOf": [
                            {
                              "type": "number"
                            },
                            {
                              "type": "null"
                            }
                          ]
                        },
                        "rawgRatingTop": {
                          "anyOf": [
                            {
                              "type": "number"
                            },
                            {
                              "type": "null"
                            }
                          ]
                        },
                        "description": {
                          "anyOf": [
                            {
                              "type": "string"
                            },
                            {
                              "type": "null"
                            }
                          ]
                        },
                        "genres": {
                          "type": "array",
                          "items": {
                            "type": "string"
                          }
                        },
                        "developers": {
                          "type": "array",
                          "items": {
                            "type": "string"
                          }
                        },
                        "publishers": {
                          "type": "array",
                          "items": {
                            "type": "string"
                          }
                        },
                        "websiteUrl": {
                          "anyOf": [
                            {
                              "type": "string"
                            },
                            {
                              "type": "null"
                            }
                          ]
                        },
                        "releases": {
                          "type": "array",
                          "items": {
                            "type": "object",
                            "required": [
                              "platformId",
                              "platformName",
                              "releaseDate",
                              "releaseDatePrecision"
                            ],
                            "properties": {
                              "platformId": {
                                "type": "string"
                              },
                              "platformName": {
                                "type": "string"
                              },
                              "releaseDate": {
                                "anyOf": [
                                  {
                                    "type": "string"
                                  },
                                  {
                                    "type": "null"
                                  }
                                ]
                              },
                              "releaseDatePrecision": {
                                "anyOf": [
                                  {
                                    "type": "string",
                                    "enum": [
                                      "day"
                                    ]
                                  },
                                  {
                                    "type": "string",
                                    "enum": [
                                      "month"
                                    ]
                                  },
                                  {
                                    "type": "string",
                                    "enum": [
                                      "year"
                                    ]
                                  },
                                  {
                                    "type": "string",
                                    "enum": [
                                      "unknown"
                                    ]
                                  }
                                ]
                              }
                            }
                          }
                        }
                      }
                    },
                    "isInWatchlist": {
                      "type": "boolean"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "error"
                  ],
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "error"
                  ],
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "error"
                  ],
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "error"
                  ],
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "servers": [
    {
      "url": "http://localhost:3001",
      "description": "Local API"
    }
  ]
}